Move pc search code to a singular services directory
