+++ js/search/components/SearchWrapper.jsx
@@ -3,7 +3,7 @@ import isArray from 'lodash.isarray';
 import isEqual from 'lodash.isequal';
 import queryString from 'query-string';
 import {search, datasources} from 'pathway-commons';
-import {queryFetch} from '../helpers/queryFetch.js';
+import {queryFetch} from '../../services/pcQueryFetch.js';
 import {Spinner} from '../../components/Spinner.jsx';
 
 // SearchWrapper
